openAI/chatagpt-basico/pyproject.toml
[tool.poetry]
name = "chatagpt-basico"
version = "0.1.0"
description = ""
authors = ["maryson011 <maryson011@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.54.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
